Read whole data from MTP device even on writing error.
Once the object bytes is requested on the MTP client device, the device tries to
send whole data of object. We need to read the complete data from the device
even when we have errors at the destination file descriptor. Otherwise the
object data will be received as a response of next request unintentionally.

BUG=23264575

Change-Id: I3369786370022f65aa760dd6b75204a946f712af